Improved code re-use across files, updated rvv compilation step
