+++ DBI_AI_Calculator.py
@@ -821,227 +821,6 @@ def printDynamoRIOARM():
             print(f"{data:12} : {opcode}")
     print("------------------------------")
 
-def parse_title_line(line):
-    parts = line.split()
-    title = {
-        "name": parts[0],
-        "isa": parts[2],
-        "precision": parts[3],
-        "threads": int(parts[4]),
-        "load": int(parts[6]),
-        "store": int(parts[8]),
-        "inst": parts[10]
-    }
-    return title
-
-#Read CARM results data
-def read_roofline_data(filename):
-    title = {}
-    data = {}
-    data_cycles = {}
-
-    with open(filename, 'r') as file:
-        title_line = file.readline().strip()
-        title = parse_title_line(title_line)
-
-        for line in file:
-            if ':' not in line:
-                continue
-
-            label, value = line.strip().split(': ')
-
-            if label == 'L1':
-                data["L1"] = float(value)
-            elif label == 'L2':
-                data["L2"] = float(value)
-            elif label == 'L3':
-                data["L3"] = float(value)
-            elif label == 'DRAM':
-                data["DRAM"] = float(value)
-            elif label == 'FP':
-                data["FP"] = float(value)
-            elif label == 'FP_FMA':
-                data["FP_FMA"] = float(value)
-            elif label == 'L1 Instruction Per Cycle':
-                data_cycles["L1"] = float(value)
-            elif label == 'L2 Instruction Per Cycle':
-                data_cycles["L2"] = float(value)
-            elif label == 'L3 Instruction Per Cycle':
-                data_cycles["L3"] = float(value)
-            elif label == 'DRAM Instruction Per Cycle':
-                data_cycles["DRAM"] = float(value)
-            elif label == 'FP Instruction Per Cycle':
-                data_cycles["FP"] = float(value)
-            elif label == 'FP_FMA Instruction Per Cycle':
-                data_cycles["FP_FMA"] = float(value)
-
-    return title, data, data_cycles
-
-
-def read_data_from_files(directory, autochoice):
-    #List all files with .out extension in the directory
-    files = [file for file in os.listdir(directory) if file.endswith('.out') and 'roofline' in file]
-
-    #Print the list of files with their indices
-    print("Available files:")
-    for i, file in enumerate(files):
-        print(f"{i + 1}. {file}")
-
-    #Prompt the user to select a file by its index
-    if autochoice == 0:
-        while True:
-            try:
-                choice = int(input("Enter the number corresponding to the file you want to read (or 0 to quit): "))
-                if 1 <= choice <= len(files):
-                    selected_file = files[choice - 1]
-                    break
-                elif choice == 0:
-                    print("No Roofline will be drawn, terminating program.")
-                    sys.exit(1)
-                else:
-                    print("Invalid choice. Please enter a number within the range.")
-            except EOFError:
-                print("No automatic roofline choice provided, no Roofline will be drawn. Exiting program.")
-                sys.exit(1)
-            except ValueError:
-                print("Invalid input. Please enter a number.")
-    else:
-        choice = autochoice
-        if 1 <= choice <= len(files):
-            selected_file = files[choice - 1]
-        else:
-            print("Autochoice does not match any file. No Roofline will be drawn, terminating program.")
-            sys.exit(1)
-
-    return read_roofline_data(os.path.join(directory, selected_file))
-
-def carm_eq(ai, bw, fp):
-    return np.minimum(ai*bw, fp)
-
-def round_power_of_2(number):
-    if number > 1:
-        for i in range(1, int(number)):
-            if (2 ** i > number):
-                return 2 ** i
-    else:
-        return 1
-
-def plot_roofline_with_dot(executable_path, exec_flops, exec_ai, choice, roi, date):
-
-    title = {}
-    data = {}
-    data_cycles = {}
-
-    executable_name = os.path.basename(executable_path)
-
-    min_ai = 0.015625
-    min_flops = 0.25
-    while exec_ai < min_ai :
-        min_ai /= 2
-    
-    while exec_flops < min_flops :
-        min_flops /= 2
-    
-    script_dir = os.path.dirname(os.path.abspath(__file__))
-
-    #Construct the path to the build folder
-    result_dir = os.path.join(script_dir, "carm_results/Roofline")
-
-    title, data, data_cycles = read_data_from_files(result_dir, choice)
-
-    fig, ax = plt.subplots(figsize=(7.1875*1.5,3.75*1.5))
-    plt.xlim(min_ai, 256)
-    plt.ylim(min_flops, round_power_of_2(int(data["FP_FMA"])))
-    ai = np.linspace(0.00390625, 256, num=200000)
-
-    #Ploting Lines
-    if title["inst"] == "fma":
-        plt.plot(ai, carm_eq(ai, data["L1"], data["FP"]), 'k', lw = 3, label='L1')
-        plt.plot(ai, carm_eq(ai, data["L2"], data["FP"]), 'grey', lw = 3, label='L2')
-        plt.plot(ai, carm_eq(ai, data["L3"], data["FP"]), 'k', linestyle='dashed', lw = 3, label='L3')
-        plt.plot(ai, carm_eq(ai, data["DRAM"], data["FP"]), 'k', linestyle='dotted', lw = 3, label='DRAM')
-    else:
-        plt.plot(ai, carm_eq(ai, data["L1"], data["FP_FMA"]), 'k', lw = 3, label='L1')
-        plt.plot(ai, carm_eq(ai, data["L2"], data["FP_FMA"]), 'grey', lw = 3, label='L2')
-        plt.plot(ai, carm_eq(ai, data["L3"], data["FP_FMA"]), 'k', linestyle='dashed', lw = 3, label='L3')
-        plt.plot(ai, carm_eq(ai, data["DRAM"], data["FP_FMA"]), 'k', linestyle='dotted', lw = 3, label='DRAM')
-        plt.plot(ai, carm_eq(ai, data["L1"], data["FP"]), 'k', linestyle='dashdot', lw = 3, label=title["inst"])
-    
-    #Plot dot at exec_gflops and exec_ai
-    plt.scatter(exec_ai, exec_flops, color='red', label=executable_name, zorder=5)
-
-    ax.spines['right'].set_visible(False)
-    ax.spines['top'].set_visible(False)
-    if roi:
-        plt.title(executable_name + " (" + title["name"] + ")" + ' DBI ROI CARM: ' + str(title["isa"]) + " " + str(title["precision"]) + " " + str(title["threads"]) + " Threads " + str(title["load"]) + " Load " + str(title["store"]) + " Store " + title["inst"], fontsize=18)
-    else:
-        plt.title(executable_name + " (" + title["name"] + ")" + ' DBI CARM: ' + str(title["isa"]) + " " + str(title["precision"]) + " " + str(title["threads"]) + " Threads " + str(title["load"]) + " Load " + str(title["store"]) + " Store " + title["inst"], fontsize=18)
-
-
-    plt.ylabel('Performance [GFLOPS/s]', fontsize=18)
-    plt.xlabel('Arithmetic Intensity [flops/bytes]', fontsize=18)
-    plt.setp(ax.get_xticklabels(), fontsize=18)
-    plt.setp(ax.get_yticklabels(), fontsize=18)
-    plt.yscale('log', base=2)
-    plt.xscale('log', base=2)
-    plt.legend(fontsize=18, loc='lower right')
-    new_rc_params = {'text.usetex': False,"svg.fonttype": 'none'}
-    plt.rcParams.update(new_rc_params)
-    plt.tight_layout()
-    if(os.path.isdir('carm_results') == False):
-            os.mkdir('carm_results')
-    if(os.path.isdir('carm_results/applications') == False):
-        os.mkdir('carm_results/applications')
-
-    if roi:
-        plt.savefig('carm_results/applications/' + executable_name + "_" + title["name"] + '_DBI_ROI_roofline_analysis_' + date + '_' + str(title["isa"]) + "_" + str(title["precision"]) + "_" + str(title["threads"]) + "_Threads_" + str(title["load"]) + "Load_" + str(title["store"]) + "Store_" + title["inst"] + '.svg')
-    else:
-        plt.savefig('carm_results/applications/' + executable_name + "_" + title["name"] + '_DBI_roofline_analysis_' + date + '_' + str(title["isa"]) + "_" + str(title["precision"]) + "_" + str(title["threads"]) + "_Threads_" + str(title["load"]) + "Load_" + str(title["store"]) + "Store_" + title["inst"] + '.svg')
-
-
-def update_csv(machine, executable_path, exec_flops, exec_ai, bandwidth, time, name, date, isa, precision, threads, method, VLEN, LMUL):
-
-    csv_path = f"./carm_results/applications/{machine}_applications.csv"
-
-    if name == "":
-        name = os.path.basename(executable_path)
-
-    if(os.path.isdir('carm_results') == False):
-        os.mkdir('carm_results')
-    if(os.path.isdir('carm_results/applications') == False):
-        os.mkdir('carm_results/applications')
-    
-    if (isa in ["rvv0.7", "rvv1.0"]):
-        isa = str(isa) + "_vl" + str(VLEN) + "_lmul" + str(LMUL)
-
-    results = [
-        date,
-        method,
-        name,
-        isa,
-        precision,
-        threads,
-        ut.custom_round(exec_ai),
-        ut.custom_round(exec_flops),
-        ut.custom_round(bandwidth),
-        ut.custom_round(time)
-    ]
-
-    headers = ['Date', 'Method', 'Name', 'ISA', 'Precision', 'Threads', 'AI', 'Gflops', 'Bandwidth', 'Time']
-
-    #Check if the file exists
-    if os.path.exists(csv_path):
-        #If exists, append without header
-        with open(csv_path, 'a', newline='') as csvfile:
-            writer = csv.writer(csvfile)
-            writer.writerow(results)
-    else:
-        #If not, write with header and include secondary headers
-        with open(csv_path, 'w', newline='') as csvfile:
-            writer = csv.writer(csvfile)
-            writer.writerow(headers)
-            writer.writerow(results)
-
 if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description="Run an executable with DynamoRIO.")
@@ -1105,8 +884,6 @@ if __name__ == "__main__":
         if args.roi:
             method += "-ROI"
 
-
-
     time_taken_seconds = float (exec_time / 1e9)
 
     flops = fp_ops/time_taken_seconds
@@ -1135,10 +912,10 @@ if __name__ == "__main__":
         if not plot_numpy == None:
             print("Manual application plotting not implemented iet, results can be viewed using the GUI")
             #date = ct.strftime('%Y-%m-%d_%H-%M-%S')
-            #plot_roofline_with_dot(args.executable_path, gflops, ai, args.choice, args.roi, date)
+            #ut.plot_roofline_with_dot(args.executable_path, gflops, ai, args.choice, args.roi, date, "dbi")
         else:
             print("No Matplotlib and/or Numpy found, in order to draw CARM graphs make sure to install them.")
     date = ct.strftime('%Y-%m-%d %H:%M:%S')
     
 
-    update_csv(args.name, args.executable_path, gflops, ai, bandwidth, time_taken_seconds, args.app_name, date, args.isa, args.precision, args.threads, method, 1, 1)
+    ut.update_csv(args.name, args.executable_path, gflops, ai, bandwidth, time_taken_seconds, args.app_name, date, args.isa, args.precision, args.threads, method, 1, 1)
